refactor(suite): split earn modal components and inline spacings
